🎨 style: refactor Account component, extract DisplayUsernameMessages, and remove redundant border classes for cleaner layout
