+++ client/src/components/Nav/SettingsTabs/Beta/Beta.tsx
@@ -4,7 +4,7 @@ import CodeArtifacts from './CodeArtifacts';
 function Beta() {
   return (
     <div className="flex flex-col gap-3 p-1 text-sm text-text-primary">
-      <div className="pb-3 last-of-type:border-b-0">
+      <div className="pb-3">
         <CodeArtifacts />
       </div>
     </div>
