+++ client/src/components/Nav/SettingsTabs/Chat/Chat.tsx
@@ -11,26 +11,26 @@ import SaveDraft from './SaveDraft';
 function Chat() {
   return (
     <div className="flex flex-col gap-3 p-1 text-sm text-text-primary">
-      <div className="pb-3 last-of-type:border-b-0">
+      <div className="pb-3">
         <FontSizeSelector />
       </div>
-      <div className="pb-3 last-of-type:border-b-0">
+      <div className="pb-3">
         <ChatDirection />
       </div>
-      <div className="pb-3 last-of-type:border-b-0">
+      <div className="pb-3">
         <SendMessageKeyEnter />
       </div>
-      <div className="pb-3 last-of-type:border-b-0">
+      <div className="pb-3">
         <ShowCodeSwitch />
       </div>
-      <div className="pb-3 last-of-type:border-b-0">
+      <div className="pb-3">
         <SaveDraft />
       </div>
       <ForkSettings />
-      <div className="pb-3 last-of-type:border-b-0">
+      <div className="pb-3">
         <ModularChat />
       </div>
-      <div className="pb-3 last-of-type:border-b-0">
+      <div className="pb-3">
         <LaTeXParsing />
       </div>
     </div>
