+++ client/src/components/Nav/SettingsTabs/General/General.tsx
@@ -147,22 +147,22 @@ function General() {
 
   return (
     <div className="flex flex-col gap-3 p-1 text-sm text-text-primary">
-      <div className="pb-3 last-of-type:border-b-0">
+      <div className="pb-3">
         <ThemeSelector theme={theme} onChange={changeTheme} />
       </div>
-      <div className="pb-3 last-of-type:border-b-0">
+      <div className="pb-3">
         <LangSelector langcode={langcode} onChange={changeLang} />
       </div>
-      <div className="pb-3 last-of-type:border-b-0">
+      <div className="pb-3">
         <UserMsgMarkdownSwitch />
       </div>
-      <div className="pb-3 last-of-type:border-b-0">
+      <div className="pb-3">
         <AutoScrollSwitch />
       </div>
-      <div className="pb-3 last-of-type:border-b-0">
+      <div className="pb-3">
         <HideSidePanelSwitch />
       </div>
-      <div className="pb-3 last-of-type:border-b-0">
+      <div className="pb-3">
         <ArchivedChats />
       </div>
     </div>
